+++ source/converter/interface_pandoc.py
@@ -1,4 +1,5 @@
 #!/usr/local/bin/python3
+
 __author__ = 'Juho Vepsäläinen, apaeffgen'
 # -*- coding: utf-8 -*-
 # The original file is modified from the following source:
@@ -27,18 +28,13 @@ __author__ = 'Juho Vepsäläinen, apaeffgen'
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-import fnmatch
-import glob
 import os
 import platform
 import subprocess
 
 from PyQt5.QtCore import QSettings
 
-
-
 from source.language.messages import *
-
 
 global fromFormat
 
@@ -256,61 +252,6 @@ def get_pandoc_options():
 
         except OSError:
             error_converter_path()
-
-
-
-
-
-def create_filelist(directory):
-
-    settings = QSettings('Pandoc', 'PanConvert')
-    filefilter = settings.value('batch_convert_filter','')
-
-    matches = []
-    for root, dirnames, filenames in os.walk(directory):
-        for filename in fnmatch.filter(filenames, '*.*'):
-            if filename != '.DS_Store':
-                matches.append(os.path.join(root, filename))
-
-    filter = filefilter.split(';')
-    matching = []
-
-    for filteritem in filter:
-
-        matching_filter = [s for s in matches if filteritem in s]
-        for i in matching_filter:
-            matching.append(i)
-
-    if len(matching) == 0:
-
-        error_file_selection()
-
-
-
-
-    return matching
-
-def create_simplefilelist():
-    settings = QSettings('Pandoc', 'PanConvert')
-    batch_settings = QSettings('Pandoc', 'PanConvert')
-    batch_open_path = batch_settings.value('batch_open_path')
-    filefilter = settings.value('batch_convert_filter','')
-
-    filelist = glob.glob(batch_open_path + '/*')
-    filter = filefilter.split(';')
-
-    matching = []
-
-    for filteritem in filter:
-
-        matching_filter = [s for s in filelist if filteritem in s]
-        for i in matching_filter:
-            matching.append(i)
-    if len(matching) == 0:
-
-        error_file_selection()
-
-    return matching
 
 
 def convert(source, to, format=None, extra_args=(), encoding='utf-8'):